+++ backend/src/main/java/iuh/fit/backend/dto/FavoriteListResponse.java
@@ -1,5 +1,7 @@
 package iuh.fit.backend.dto;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -14,6 +16,6 @@ public class FavoriteListResponse {
     private String productName;
     private String productSlug;
     private String productImageUrl;
-    private Double productPrice;
-    private Double productSalePrice;
+    private BigDecimal productPrice;
+    private BigDecimal productSalePrice;
 }
